diff: rename differ to comparer
Remove combined interface and split implementations.
