+++ cmd/containerd/builtins.go
@@ -2,7 +2,7 @@ package main
 
 // register containerd builtins here
 import (
-	_ "github.com/containerd/containerd/diff/walking"
+	_ "github.com/containerd/containerd/diff/walking/plugin"
 	_ "github.com/containerd/containerd/gc/scheduler"
 	_ "github.com/containerd/containerd/services/containers"
 	_ "github.com/containerd/containerd/services/content"
