+++ rootfs/diff.go
@@ -10,11 +10,11 @@ import (
 	"golang.org/x/net/context"
 )
 
-// Diff creates a layer diff for the given snapshot identifier from the parent
-// of the snapshot. A content ref is provided to track the progress of the
-// content creation and the provided snapshotter and mount differ are used
+// CreateDiff creates a layer diff for the given snapshot identifier from the
+// parent of the snapshot. A content ref is provided to track the progress of
+// the content creation and the provided snapshotter and mount differ are used
 // for calculating the diff. The descriptor for the layer diff is returned.
-func Diff(ctx context.Context, snapshotID string, sn snapshots.Snapshotter, d diff.Differ, opts ...diff.Opt) (ocispec.Descriptor, error) {
+func CreateDiff(ctx context.Context, snapshotID string, sn snapshots.Snapshotter, d diff.Comparer, opts ...diff.Opt) (ocispec.Descriptor, error) {
 	info, err := sn.Stat(ctx, snapshotID)
 	if err != nil {
 		return ocispec.Descriptor{}, err
@@ -42,5 +42,5 @@ func Diff(ctx context.Context, snapshotID string, sn snapshots.Snapshotter, d di
 		defer sn.Remove(ctx, upperKey)
 	}
 
-	return d.Diff(ctx, lower, upper, opts...)
+	return d.Compare(ctx, lower, upper, opts...)
 }
